pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-public-health"
version = "0.1.0"
description = "RAG pipeline for Principles of Public Health PDF"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "PyMuPDF>=1.23",
    "Pillow>=10.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "openai>=1.0",
    "click>=8.0",
    "rich>=13.0",
    "numpy>=1.24",
    "ragas>=0.4.3",
    "streamlit>=1.30",
    "langfuse>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
ocr-extra = [
    "easyocr>=1.7",
    "paddleocr>=2.7",
    "paddlepaddle>=2.5",
]
vlm = [
    "transformers>=4.40",
    "accelerate>=0.27",
    "qwen-vl-utils>=0.0.8",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
